requirements.txt
promptlayer==1.0.28
python-dotenv==1.0.1
matplotlib==3.9.2
openai==1.53.0